chore(deps-dev): revert back to surge 0.23.1
This project includes a patch for version 0.23.1, and upgrading the version of surge require to
investigate if this patch is still needed.
Depending on the result of the investigation, this patch will be adapted or removed.

Currently, applying the patch to 0.24.x generates an error.
So reverting until we conduct the investigation.

This reverts commit f3e7fa34319940c7955cbf57fd558408e1adfb5b.